Add precise dep tracking and data relocations
This is a large rework in an attempt to cleanly separate relocs
from module splitting.
A bunch of things, like export names and exact imported items
are now decided before they are emitted.

DepNodes are the central vehicle for analysis. They are used to
decide what items from the main wasm we include, import and
export from each module. The function-based usage is expanded
to cover all types of items in a wasm, except custom sections.

Globals, tables, tags and memories are shared via name. That is,
they get an appropriate export in the defining module and an
import in the consuming module (in the __wasm_split namespace).
Currently, all such items are export from the main module in any case.

Functions are shared via an entry in the indirect function table, the
assignment of which is decided in emit::IndirectFunctionEmitInfo.
They do not need an explicit export/import.

Data items are shared via memory address. To compact these, defined
symbols are relocated shuffled in each data segment. The assignment
of their final address happens in emit::DataEmitInfo.