generate a unique name for the marker symbol
technically, there is no reason for this symbol to be public, but
the compiler doesn't seem to respect the `#[link_section]` on
version marker otherwise (see linked bug in rust-lang/rust).
To not collide with other symbols/library versions, the name is
now derived from wasm_split_helper's manifest path.