Stricter handling of data dependencies (#19)
* be stricter in the exported names

we at least check for duplicates and use a consistent function to refer
to the dep nodes for the main memory and indirect function table.
removes the "simplified" export name for the stack pointer, so we don't
have to parse for it when reading symbols

* optimize searching for the function getting targeted by a relocation

use an exponential search strategy instead of always binary searching.
Since we iterate over all (code) relocations, this reduces the time spent
in searching (for large number of functions)

* smarter data dependency resolution

by recording dependencies between overlapping data items,
we prevent duplication of some items. The implementation might seem
a bit dense. Read carefully!

* print file sizes in the test runner

* only do major version patches in the library part

for one, we are less strict in the library dependencies of the cli.
Version requirements in libs should be as lenient as possible to allow
for more consumers (specifically, regex 1.12 is not
would use 'increase-if-necessary' as a strategy, but that is not supported
(see dependabot/dependabot-core#4009)